Change file_size type to long
The next Bot API update will change file_size properties from int to long to accomodate 4GB+ files

+++ src/Telegram.Bot/Types/FileBase.cs
@@ -29,5 +29,5 @@ public abstract class FileBase
     /// Optional. File size
     /// </summary>
     [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-    public int? FileSize { get; set; }
+    public long? FileSize { get; set; }
 }
